app/llm/requirements.txt
# PDF 처리 관련
PyPDF2==3.0.1
pdfplumber==0.10.3
camelot-py==0.11.0
tabula-py==2.9.0
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# 임베딩 및 벡터 DB 관련
chromadb==0.4.22
openai==1.12.0

# RAG 및 LLM 관련
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.13

# 다국어 처리
langdetect==1.0.9
googletrans==4.0.0-rc1

# 유틸리티
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
numpy==1.26.3
pandas==2.1.4

# 테스트
pytest==7.4.3

# 웹 UI (추가)
streamlit==1.32.0
matplotlib==3.8.2
plotly==5.18.0

# 음성 처리 (추가)
pyaudio==0.2.13
SpeechRecognition==3.10.0
pydub==0.25.1
